+++ src/main/java/umc/spring/service/MissionService/MissionQueryService.java
@@ -1,0 +1,4 @@
+package umc.spring.service.MissionService;
+
+public interface MissionQueryService {
+}
