+++ src/main/java/umc/spring/service/ReviewService/ReviewQueryService.java
@@ -1,0 +1,4 @@
+package umc.spring.service.ReviewService;
+
+public interface ReviewQueryService {
+}
