+++ src/main/java/umc/spring/service/ReviewService/ReviewQueryServiceImpl.java
@@ -1,9 +1,9 @@
-package umc.spring.service.MemberService;
+package umc.spring.service.ReviewService;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MemberQueryServiceImpl implements MemberQueryService{
+public class ReviewQueryServiceImpl implements ReviewQueryService{
 }
