+++ src/main/java/umc/spring/service/StoreService/StoreQueryService.java
@@ -1,0 +1,4 @@
+package umc.spring.service.StoreService;
+
+public interface StoreQueryService {
+}
